lots of changes made
There have been a lot of changes to the code. I ran the code through an html validator and there was a lot of errors which i have corrected. The html structure is now correct. I have also added some css to make the webpage look user friendly for the moment.